requirements.txt
# YAML
pyyaml

# Numbers to text
num2words

# HTML template
Jinja2

# PDF (also requires wkhtmltopdf)
pdfkit

# LB exchange rates
suds

# Colorful logging
colorlog

# Currency formatting
money
babel

# Unit tests + coverage
nose
coverage
